Fix leadership lost just before the WalFrames replication hook
When the leader loses leadership just before running the WalFrames
replication hook (typically to commit a transaction), it now properly
marks the transaction as stale (so the automatic WAL rollback
initiated by SQLite will effectively be a no-op) and start a surrogate
follower (so the next leader can roll back the leftover transaction
that it finds).

Also, the demo program has been changed to handle tx.Commit() failures
differently: in that case there's no need to call tx.Rollback(), as
that is automatically handled by SQLite.

Fixes #2.