Close the connection stream handle after the timer one
The reason is that the close callback for the stream handle also releases the
memory of the connection object, which holds the timer handle as well. Since
libuv has some handle-involving logic triggering after uv_close(), even in the
case where the close callback is set to NULL, the previous code could result in
invalid memory accesses.