Manually fire raft apply callback when closing gateway
If the connection gets closed while a raft apply request is in progress, we now
manually trigger the raft apply request callback so that SQLite can unwind its
state and we can finalize the statement that was being executed. When the real
raft apply request callback fires, we just ignore the result.