leader: remove raft_apply member from struct leader
When the gateway is closed, and the leader is freed, it's possible that
there is a still a reference to the raft_apply member in the leader
requests queue of the raft node. Later, when raft tries to access that
memory when cleaning up its leader requests queue a memory fault can be
triggered.

Solve this by allocating the request and only freeing it in the callback.