db/leader: Cleanup after sqlite3_open_v2
From the SQLite docs on sqlite3_open_v2
A database connection handle is usually returned in *ppDb, even if an
error occurs. The only exception is that if SQLite is unable to allocate
memory to hold the sqlite3 object, a NULL will be written into *ppDb
instead of a pointer to the sqlite3 object.

Whether or not an error occurs when it is opened, resources associated
with the database connection handle should be released by passing it to
sqlite3_close() when it is no longer required.