Separate parameter parsing and binding
Rationale: we want client request parsing to happen on the main thread,
and we want everything beginning with sqlite3_prepare_v2 to happen on
the DB thread (since the DB will have the database open with the
appropriate VFS).

There is an unresolved issue that the parsed statement parameters borrow
from the gateway's request buffer, so we need to ensure that the DB
thread finishes its work before that buffer ends its lifetime.

Signed-off-by: Cole Miller <cole.miller@canonical.com>