server: Replace monitor_cb by state_cb.
The state_cb is triggered immediately when the raft state changes,
instead of the monitor_cb, that is only polling for the state change at
the start of a loop iteration. This avoids some bugs where dqlite has to
perform cleanups on leadership loss, but wasn't informed yet of the
leadership loss.

Signed-off-by: Mathieu Borderé <mathieu.bordere@canonical.com>