Replace vfs2_wal_frame with dqlite_vfs_frame
vfs2_wal_frame carries the commit marker explicitly, so it doesn't match
COMMAND_FRAMES, which only carries the page numbers and page data.
dqlite_vfs_frame is defined with this in mind. To get the commit markers
on the follower side, we start with the size of the main file in pages
and take the maximum with each page number up to the commit frame in the
WAL. This matches what the in-memory VFS does.

Signed-off-by: Cole Miller <cole.miller@canonical.com>