Move shm finalization to maybe_close_entry
This ties the destruction of the shared memory to that of the owning
entry, because they are created together. If shm finalization is tied to
a call to xShmUnmap, then it can't happen unless SQLite called xShmMap
in the first place, and this is not guaranteed to happen. For example,
if we start and stop a follower node without ever running a checkpoint
or transaction, SQLite will never get the change to call xShmMap.

Signed-off-by: Cole Miller <cole.miller@canonical.com>