Simplify and correct W-i hdr cksum calculations
First, there's no need for the endianness parameter to the checksum
calculation unless we're copying the contents of the database directory
between machines with different endiannesses. This was not properly
supported by the old code (we just used native_magic() everywhere), and
it seems like an edge case, so let's cut the complexity for now.

Second, factor out the repeated logic for calculating and writing the
checksums in the WAL-index header. This fixes a bug that showed up in
vfs2_apply where we were checksumming the wrong number of bytes.

Signed-off-by: Cole Miller <cole.miller@canonical.com>