Factor out shared memory into a struct
Grouping together the fields that relate to the shm into a discrete type
makes it easier to read and reason about the code that touches these
fields.

In the process, this commit also changes the memory allocation strategy
for the shm. Instead of having one allocation per shm region and another
allocation to hold the array of pointers, we have one big allocation
that contains all the existing regions, one after the other. This cuts
down significantly on malloc and free calls.

We also implement proper management of shm hash tables beyond the first.

Signed-off-by: Cole Miller <cole.miller@canonical.com>