Split the WAL swap and the WAL header write
This cleaves the dqlite-specific WAL swap, which targets the volatile
state, from writing a new WAL header, which is pure SQLite and targets
the persistent state. It's not a completely clean split because the new
wal_swap function does write the outgoing WAL and flip the moving name,
but I think it's still a win for readability. This change also supports
an upcoming refactor to vfs2_add_uncommitted.

Signed-off-by: Cole Miller <cole.miller@canonical.com>