Add some ancillary state machines for observability
We'd like to make diagrams showing what's going on inside vfs2 in more
detail and how it relates to the "big" entry state machine. This commit
enables that by introducing new state machines that model four bits of
state: the write lock, the recovery lock, the checkpoint lock, and the
shared memory. The invariants are vacuous, and even the allowed
transitions aren't very constraining, but this is quite handy for
understanding what's going on e.g. during tests.

Signed-off-by: Cole Miller <cole.miller@canonical.com>