threadpool: Initialize async handle last
It's unsound to uv_close the async handle and then immediately free the
pool_impl_t, because libuv will still hold a pointer to the handle.
Avoid the need to do this by calling uv_async_init last, so there's no
need to roll it back when a later operation fails.

Signed-off-by: Cole Miller <cole.miller@canonical.com>