Propagate raft_recover errors and improve tracing
At the moment, the raft_recover error messages are not propagated,
so the caller gets a cryptic error code (1), without further
details.

This change ensures that the error messages are propagated and adds
a few extra trace messages.

Furthermore, tracing is disabled unless we start the node (not the
case when we're actually trying to recover it). For now, we're adding a
"dqliteTracingMaybeEnable" call in "dqlite_node_recover_ext".