feat: use SQLITE_STATIC binding
The whole protocol and the `gateway.c` implementation assume that the
current request message is fixed in memory and is never overwritten
until a response has been sent (including multiple message responses
like the query rows). This is also why the part around interrupting long
queries doesn't work.

Given that the solution to the above would be to allocate a message for
each request, it is safe to use `SQLITE_STATIC` and let SQLite reference
data in the message.